update: renamed TabsService file name
