+++ src/app/services/index.ts
@@ -1,6 +1,6 @@
 export { EmitHistoryService } from './emit-history.service';
 export { ProgressBarService } from './progress-bar.service';
 export { EmitterService } from './emitter.service';
-export { EmitterTabsService } from './tabs.service';
+export { EmitterTabsService } from './emitter-tabs.service';
 export { FileService } from './file.service';
 export { StateService } from './state.service';
